requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.3.1
    # via kombu
asgiref==3.8.1
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
    #   django-minify-html
attrs==25.3.0
    # via
    #   service-identity
    #   twisted
autobahn==24.4.2
    # via daphne
automat==24.8.1
    # via twisted
billiard==4.2.1
    # via celery
boto3==1.35.29
    # via -r requirements.in
botocore==1.35.99
    # via
    #   boto3
    #   s3transfer
celery==5.4.0
    # via
    #   -r requirements.in
    #   django-celery-beat
    #   flower
certifi==2025.1.31
    # via
    #   requests
    #   sentry-sdk
cffi==1.17.1
    # via cryptography
channels==4.1.0
    # via
    #   -r requirements.in
    #   channels-redis
channels-redis==4.2.0
    # via -r requirements.in
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
constantly==23.10.4
    # via twisted
crispy-bootstrap4==2024.1
    # via -r requirements.in
cron-descriptor==1.4.5
    # via django-celery-beat
cryptography==44.0.2
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==4.1.2
    # via -r requirements.in
django==5.1.1
    # via
    #   -r requirements.in
    #   channels
    #   crispy-bootstrap4
    #   django-allauth
    #   django-appconf
    #   django-cachalot
    #   django-celery-beat
    #   django-crispy-forms
    #   django-filter
    #   django-hcaptcha-field
    #   django-minify-html
    #   django-redis
    #   django-select2
    #   django-sql-utils
    #   django-storages
    #   django-taggit
    #   django-timezone-field
    #   django-tinymce
    #   sentry-sdk
django-allauth==65.0.2
    # via -r requirements.in
django-appconf==1.1.0
    # via django-select2
django-cachalot==2.6.3
    # via -r requirements.in
django-celery-beat==2.8.1
    # via -r requirements.in
django-crispy-forms==2.3
    # via
    #   -r requirements.in
    #   crispy-bootstrap4
django-filter==24.3
    # via -r requirements.in
django-hcaptcha-field==1.4.0
    # via -r requirements.in
django-ipware==7.0.1
    # via -r requirements.in
django-minify-html==1.9.0
    # via -r requirements.in
django-redis==5.4.0
    # via -r requirements.in
django-select2==8.2.1
    # via -r requirements.in
django-sql-utils==0.7.0
    # via -r requirements.in
django-storages==1.14.4
    # via -r requirements.in
django-taggit==6.1.0
    # via -r requirements.in
django-timezone-field==7.1
    # via django-celery-beat
django-tinymce==4.1.0
    # via -r requirements.in
flower==2.0.1
    # via -r requirements.in
humanize==4.12.1
    # via flower
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.10
    # via
    #   hyperlink
    #   requests
    #   twisted
incremental==24.7.2
    # via twisted
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.5.0
    # via celery
minify-html==0.15.0
    # via django-minify-html
msgpack==1.1.0
    # via channels-redis
nanoid==2.0.0
    # via -r requirements.in
pillow==10.4.0
    # via -r requirements.in
prometheus-client==0.21.1
    # via flower
prompt-toolkit==3.0.50
    # via click-repl
psycopg2==2.9.10
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.1
    # via service-identity
pycparser==2.22
    # via cffi
pyjwt==2.9.0
    # via -r requirements.in
pyopenssl==25.0.0
    # via twisted
python-crontab==3.2.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   celery
    #   python-crontab
python-dotenv==1.0.1
    # via -r requirements.in
python-ipware==3.0.0
    # via django-ipware
python-magic==0.4.27
    # via -r requirements.in
pytz==2025.1
    # via flower
redis==5.2.1
    # via
    #   channels-redis
    #   django-redis
requests==2.32.3
    # via -r requirements.in
s3transfer==0.10.4
    # via boto3
sentry-sdk[django]==2.27.0
    # via -r requirements.in
service-identity==24.2.0
    # via twisted
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via
    #   django
    #   django-sql-utils
tornado==6.4.2
    # via flower
twisted[tls]==24.11.0
    # via daphne
txaio==23.1.1
    # via autobahn
typing-extensions==4.12.2
    # via
    #   pyopenssl
    #   twisted
tzdata==2025.1
    # via
    #   celery
    #   django-celery-beat
    #   kombu
urllib3==2.3.0
    # via
    #   botocore
    #   requests
    #   sentry-sdk
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
zope-interface==7.2
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
